requirements.txt
absl-py==0.11.0
asgiref==3.3.0
astunparse==1.6.3
backcall==0.1.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.1.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.2.1
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
sqlparse==0.4.1
tb-nightly==2.4.0a20201101
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0rc3
tensorflow-estimator==2.2.0
termcolor==1.1.0
tf-estimator-nightly==2.4.0.dev2020102301
tf-nightly==2.2.0.dev20200428
tf-nightly-gpu==2.5.0.dev20201029
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
